Added SmartEnemy and fixed sticky keys.

+++ GameProject/src/pc/petko/main/Game.java
@@ -111,7 +111,7 @@ public class Game extends Canvas implements Runnable{
 	}
 	
 	// Prohibits box to move out of canvas
-	public static int clamp(int var, int min, int max) {
+	public static float clamp(float var, float min, float max) {
 		if (var >= max) {
 			return var = max;
 		} else if (var <= min) {
